🚚 rename: 전체 공고 관련 폴더 정리
