+++ pages/search/components/AllNotice/components/HeaderText/index.tsx
